WIP: mock loading images from URL for screenshot tests
MOB-4082
