+++ widgetssdk/src/testSnapshot/java/com/glia/widgets/chat/adapter/holder/OperatorStatusViewHolderSnapshotTest.kt
@@ -8,14 +8,14 @@ import com.glia.widgets.di.Dependencies
 import com.glia.widgets.snapshotutils.SnapshotChatScreen
 import com.glia.widgets.snapshotutils.SnapshotChatView
 import com.glia.widgets.snapshotutils.SnapshotLottie
-import com.glia.widgets.snapshotutils.SnapshotPicasso
+import com.glia.widgets.snapshotutils.SnapshotCoil
 import com.glia.widgets.snapshotutils.SnapshotProviders
 import com.glia.widgets.view.unifiedui.theme.UnifiedTheme
 import org.junit.Test
 
 internal class OperatorStatusViewHolderSnapshotTest : SnapshotTest(
     renderingMode = fullWidthRenderMode
-), SnapshotChatScreen, SnapshotChatView, SnapshotProviders, SnapshotPicasso, SnapshotLottie {
+), SnapshotChatScreen, SnapshotChatView, SnapshotProviders, SnapshotCoil, SnapshotLottie {
 
     // MARK: Connected
 
@@ -215,7 +215,7 @@ internal class OperatorStatusViewHolderSnapshotTest : SnapshotTest(
         unifiedTheme: UnifiedTheme? = null
     ): ViewData {
         lottieMock()
-        picassoMock(imageLoadError = imageLoadError)
+//        mockCoil(imageLoadError = imageLoadError) // TODO: need to make mockCoil accept optional boolean
 
         unifiedTheme?.let { Dependencies.gliaThemeManager.theme = it }
